+++ src/ProChat/container/Provider.tsx
@@ -54,16 +54,6 @@ export const ProChatProvider = memo<ProChatProviderProps<any>>(
     return (
       <Provider createStore={() => createStore(props, devtoolOptions)}>
         {Content}
-        <StoreUpdater
-          chatRef={props.chatRef}
-          init={!loading}
-          helloMessage={helloMessage}
-          chats={chats}
-          userMeta={userMeta}
-          request={request}
-          assistantMeta={assistantMeta}
-          onChatsChange={onChatsChange}
-        />
       </Provider>
     );
   },