+++ src/ProChat/store/action.ts
@@ -103,6 +103,27 @@ export interface ChatAction {
    * @returns  消息 id
    */
   getMessageId: (messages: ChatMessage[], parentId: string) => Promise<string>;
+
+  /**
+   * 用于更新一条消息的内容（目前仅用于平滑输出时候，其余情况请直接使用 dispatchMessage ）
+   */
+  updateMessageContent: (id: string, content: string) => Promise<void>;
+
+  /**
+   * 创建一条平滑输出的内容
+   */
+  createSmoothMessage: (id: string) => {
+    startAnimation: (speed?: number) => Promise<void>;
+    stopAnimation: () => void;
+    outputQueue: string[];
+    isAnimationActive: boolean;
+  };
+
+  /**
+   * 获取当前 loading 生成的消息 id
+   * @returns  消息 id ｜ undefined
+   */
+  getChatLoadingId: () => string | undefined;
 }
 
 export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], [], ChatAction> = (
@@ -124,6 +145,12 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
     get().dispatchMessage({ id, type: 'deleteMessage' });
   },
 
+  updateMessageContent: async (id, content) => {
+    const { dispatchMessage, updateMessageContent } = get();
+    dispatchMessage({ id, key: 'content', type: 'updateMessage', value: content });
+    updateMessageContent(id, content);
+  },
+
   dispatchMessage: (payload) => {
     const { chats, onChatsChange } = get();
 
@@ -134,14 +161,19 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
     onChatsChange?.(nextChats);
   },
   generateMessage: async (messages, assistantId) => {
-    const { dispatchMessage, toggleChatLoading, config, defaultModelFetcher } = get();
-
+    const {
+      dispatchMessage,
+      toggleChatLoading,
+      config,
+      defaultModelFetcher,
+      createSmoothMessage,
+      updateMessageContent,
+    } = get();
     const abortController = toggleChatLoading(
       true,
       assistantId,
       t('generateMessage(start)', { assistantId, messages }) as string,
     );
-
 
     // ========================== //
     //   对 messages 做统一预处理    //
@@ -151,8 +183,8 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
     const slicedMessages = getSlicedMessagesWithConfig(messages, config);
 
     // 2. 替换 inputMessage 模板
-    const compilerMessages =(slicedMessages:ChatMessage[])=>{
-    const compiler = template(config.inputTemplate, { interpolate: /{{([\S\s]+?)}}/g });
+    const compilerMessages = (slicedMessages: ChatMessage[]) => {
+      const compiler = template(config.inputTemplate, { interpolate: /{{([\S\s]+?)}}/g });
       return slicedMessages.map((m) => {
         if (m.role === 'user') {
           try {
@@ -165,10 +197,8 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
         }
         return m;
       });
-    }
-    const postMessages = !config.inputTemplate
-      ? slicedMessages
-      : compilerMessages(slicedMessages);
+    };
+    const postMessages = !config.inputTemplate ? slicedMessages : compilerMessages(slicedMessages);
 
     // 3. 添加 systemRole
     if (config.systemRole) {
@@ -188,27 +218,34 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
     let output = '';
     let isFunctionCall = false;
 
-    let timeId = 0;
+    const { startAnimation, stopAnimation, outputQueue, isAnimationActive } =
+      createSmoothMessage(assistantId);
 
     await fetchSSE(fetcher, {
       onErrorHandle: (error) => {
         dispatchMessage({ id: assistantId, key: 'error', type: 'updateMessage', value: error });
       },
+      onAbort: async () => {
+        stopAnimation();
+      },
+      onFinish: async (content) => {
+        stopAnimation();
+        if (outputQueue.length > 0 && !isFunctionCall) {
+          await startAnimation(15);
+        }
+        await updateMessageContent(assistantId, content);
+      },
       onMessageHandle: (text) => {
         output += text;
 
-        // 如果存在上一个定时器，那么清除
-        if (timeId) clearTimeout(timeId);
+        if (!isAnimationActive && !isFunctionCall) startAnimation();
 
-        // 使用定时器来监听性能，保证出入的时候不会太卡顿
-        timeId = (window.requestIdleCallback || window.setTimeout)(() => {
-          dispatchMessage({
-            id: assistantId,
-            key: 'content',
-            type: 'updateMessage',
-            value: output,
-          });
-        });
+        if (abortController?.signal.aborted) {
+          // aborted 后停止当前输出
+          return;
+        } else {
+          outputQueue.push(text);
+        }
 
         // TODO: need a function call judge callback
         // 如果是 function call
@@ -218,9 +255,21 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
       },
     });
 
-    (window.requestIdleCallback || window.setTimeout)(() => {
-      toggleChatLoading(false, undefined, t('generateMessage(end)') as string);
-    });
+    let timeoutId; // 用于存储轮询队列的计时器id
+    const checkAndToggleChatLoading = () => {
+      clearTimeout(timeoutId); // 清除任何现有的计时器
+      // 等待队列内容输出完毕
+      if (outputQueue === undefined || outputQueue.length === 0 || outputQueue.toString() === '') {
+        // 当队列为空时
+        toggleChatLoading(false, undefined, t('generateMessage(end)') as string);
+        clearTimeout(timeoutId);
+      } else {
+        // 如果队列不为空，则设置一个延迟或者使用某种形式的轮询来再次检查队列
+        timeoutId = setTimeout(checkAndToggleChatLoading, 30); // CHECK_INTERVAL 是毫秒数，代表检查间隔时间
+      }
+    };
+
+    checkAndToggleChatLoading();
 
     return { isFunctionCall };
   },
@@ -309,7 +358,20 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
   },
 
   stopGenerateMessage: () => {
-    const { abortController, toggleChatLoading } = get();
+    const { abortController, toggleChatLoading, chatLoadingId, chats, dispatchMessage } = get();
+    // 如果当前 最后一条为 chatLoadingId 停止前需要清空
+    if (chats && chats.length > 0) {
+      const lastChat = chats[chats.length - 1];
+      if (lastChat.content === LOADING_FLAT) {
+        dispatchMessage({
+          id: chatLoadingId,
+          key: 'content',
+          type: 'updateMessage',
+          value: '',
+        });
+      }
+    }
+
     if (!abortController) return;
 
     abortController.abort();
@@ -337,7 +399,8 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
       params,
     );
 
-    if (typeof request === 'function') return request(payload.messages as ChatMessage[], payload);
+    if (typeof request === 'function')
+      return request(payload.messages as ChatMessage[], payload, options?.signal);
 
     const url = typeof request === 'string' ? request : '/api/openai/chat';
 
@@ -352,5 +415,75 @@ export const chatAction: StateCreator<ChatStore, [['zustand/devtools', never]], 
     const { genMessageId } = get();
     if (typeof genMessageId === 'function') return genMessageId(messages, parentId);
     return nanoid();
+  },
+
+  createSmoothMessage: (id) => {
+    const { dispatchMessage } = get();
+
+    let buffer = '';
+    // why use queue: https://shareg.pt/GLBrjpK
+    let outputQueue: string[] = [];
+
+    // eslint-disable-next-line no-undef
+    let animationTimeoutId: NodeJS.Timeout | null = null;
+    let isAnimationActive = false;
+
+    // when you need to stop the animation, call this function
+    const stopAnimation = () => {
+      isAnimationActive = false;
+      if (animationTimeoutId !== null) {
+        clearTimeout(animationTimeoutId);
+        animationTimeoutId = null;
+      }
+    };
+
+    // define startAnimation function to display the text in buffer smooth
+    // when you need to start the animation, call this function
+    const startAnimation = (speed = 2) =>
+      new Promise<void>((resolve) => {
+        if (isAnimationActive) {
+          resolve();
+          return;
+        }
+
+        isAnimationActive = true;
+
+        const updateText = () => {
+          // 如果动画已经不再激活，则停止更新文本
+          if (!isAnimationActive) {
+            clearTimeout(animationTimeoutId!);
+            animationTimeoutId = null;
+            resolve();
+          }
+
+          // 如果还有文本没有显示
+          // 检查队列中是否有字符待显示
+          if (outputQueue.length > 0) {
+            // 从队列中获取前两个字符（如果存在）
+            const charsToAdd = outputQueue.splice(0, speed).join('');
+            buffer += charsToAdd;
+
+            // 更新消息内容，这里可能需要结合实际情况调整
+            dispatchMessage({ id, key: 'content', type: 'updateMessage', value: buffer });
+
+            // 设置下一个字符的延迟
+            animationTimeoutId = setTimeout(updateText, 16); // 16 毫秒的延迟模拟打字机效果
+          } else {
+            // 当所有字符都显示完毕时，清除动画状态
+            isAnimationActive = false;
+            animationTimeoutId = null;
+            resolve();
+          }
+        };
+
+        updateText();
+      });
+
+    return { startAnimation, stopAnimation, outputQueue, isAnimationActive };
+  },
+
+  getChatLoadingId: () => {
+    const { chatLoadingId } = get();
+    return chatLoadingId;
   },
 });